Refactoring show module, bump version and re-work setup.py
