+++ tests/test_show.py
@@ -5,13 +5,13 @@
 
 import json
 import mysql.connector
-import wwdtm.show as show
+from wwdtm.show import details, info, utility
 
 def test_id_exists(show_id: int,
                    database_connection: mysql.connector.connect,
                    print_response: bool = False):
     """Testing repsonse from show.id_exists"""
-    response = show.id_exists(show_id, database_connection)
+    response = utility.id_exists(show_id, database_connection)
     assert response
     if print_response:
         print(json.dumps(response, indent=2))
@@ -20,7 +20,7 @@ def test_id_not_exists(show_id: int,
                        database_connection: mysql.connector.connect,
                        print_response: bool = False):
     """Testing repsonse from show.id_exists with invalid ID"""
-    response = show.id_exists(show_id, database_connection)
+    response = utility.id_exists(show_id, database_connection)
     assert not response
     if print_response:
         print(json.dumps(response, indent=2))
@@ -31,10 +31,10 @@ def test_date_exists(show_year: int,
                      database_connection: mysql.connector.connect,
                      print_response: bool = False):
     """Testing response from show.date_exists"""
-    response = show.date_exists(show_year,
-                                show_month,
-                                show_day,
-                                database_connection)
+    response = utility.date_exists(show_year,
+                                   show_month,
+                                   show_day,
+                                   database_connection)
     assert response
     if print_response:
         print(json.dumps(response, indent=2))
@@ -45,10 +45,10 @@ def test_date_not_exists(show_year: int,
                          database_connection: mysql.connector.connect,
                          print_response: bool = False):
     """Testing response from show.date_exists with incorrect date"""
-    response = show.date_exists(show_year,
-                                show_month,
-                                show_day,
-                                database_connection)
+    response = utility.date_exists(show_year,
+                                   show_month,
+                                   show_day,
+                                   database_connection)
     assert not response
     if print_response:
         print(json.dumps(response, indent=2))
@@ -57,7 +57,7 @@ def test_retrieve_by_id(show_id: int,
                         database_connection: mysql.connector.connect,
                         print_response: bool = False):
     """Testing response from show.retrieve_by_id"""
-    show_info = show.retrieve_by_id(show_id, database_connection)
+    show_info = info.retrieve_by_id(show_id, database_connection)
     assert show_info is not None
     assert "best_of" in show_info
     assert "repeat_show" in show_info
@@ -68,7 +68,7 @@ def test_retrieve_by_invalid_id(show_id: int,
                                 database_connection: mysql.connector.connect,
                                 print_response: bool = False):
     """Testing response from show.retrieve_by_id with invalid ID"""
-    show_info = show.retrieve_by_id(show_id, database_connection)
+    show_info = info.retrieve_by_id(show_id, database_connection)
     assert show_info is None
     if print_response:
         print(json.dumps(show_info, indent=2))
@@ -79,7 +79,7 @@ def test_retrieve_by_date(show_year: int,
                           database_connection: mysql.connector.connect,
                           print_response: bool = False):
     """Testing response from show.retrieve_by_date"""
-    show_info = show.retrieve_by_date(show_year,
+    show_info = info.retrieve_by_date(show_year,
                                       show_month,
                                       show_day,
                                       database_connection)
@@ -96,7 +96,7 @@ def test_retrieve_by_invalid_date(show_year: int,
                                   print_response: bool = False):
     """Testing response from show.retrieve_by_date with incorrect date"""
     try:
-        show_info = show.retrieve_by_date(show_year,
+        show_info = info.retrieve_by_date(show_year,
                                           show_month,
                                           show_day,
                                           database_connection)
@@ -110,7 +110,7 @@ def test_retrieve_by_date_string(show_date: str,
                                  database_connection: mysql.connector.connect,
                                  print_response: bool = False):
     """Testing response from show.retrieve_by_date_string"""
-    show_info = show.retrieve_by_date_string(show_date, database_connection)
+    show_info = info.retrieve_by_date_string(show_date, database_connection)
     assert show_info is not None
     if print_response:
         print(json.dumps(show_info, indent=2))
@@ -120,7 +120,7 @@ def test_retrieve_by_invalid_date_string(show_date: str,
                                          print_response: bool = False):
     """Testing response from show.retrieve_by_date_string with incorrect date"""
     try:
-        show_info = show.retrieve_by_date_string(show_date, database_connection)
+        show_info = info.retrieve_by_date_string(show_date, database_connection)
         assert show_info is None
         if print_response:
             print(json.dumps(show_info, indent=2))
@@ -131,7 +131,7 @@ def test_retrieve_months_by_year(show_year: int,
                                  database_connection: mysql.connector.connect,
                                  print_response: bool = False):
     """Testing response from show.retrieve_months_by_year"""
-    show_months = show.retrieve_months_by_year(show_year,
+    show_months = info.retrieve_months_by_year(show_year,
                                                database_connection)
     assert show_months is not None
     if print_response:
@@ -140,7 +140,7 @@ def test_retrieve_months_by_year(show_year: int,
 def test_retrieve_years(database_connection: mysql.connector.connect,
                         print_response: bool = False):
     """Testing response from show.retrieve_years"""
-    show_years = show.retrieve_years(database_connection)
+    show_years = info.retrieve_years(database_connection)
     assert show_years is not None
     if print_response:
         print(json.dumps(show_years, indent=2))
@@ -149,7 +149,7 @@ def test_retrieve_by_year(show_year: int,
                           database_connection: mysql.connector.connect,
                           print_response: bool = False):
     """Testing response form show.retrieve_by_year"""
-    show_info = show.retrieve_by_year(show_year, database_connection)
+    show_info = info.retrieve_by_year(show_year, database_connection)
     assert show_info is not None
     if print_response:
         print(json.dumps(show_info, indent=2))
@@ -159,7 +159,7 @@ def test_retrieve_by_year_month(show_year: int,
                                 database_connection: mysql.connector.connect,
                                 print_response: bool = False):
     """Testing response form show.retrieve_by_year_month"""
-    show_info = show.retrieve_by_year_month(show_year,
+    show_info = info.retrieve_by_year_month(show_year,
                                             show_month,
                                             database_connection)
     assert show_info is not None
@@ -169,7 +169,7 @@ def test_retrieve_by_year_month(show_year: int,
 def test_retrieve_all(database_connection: mysql.connector.connect,
                       print_response: bool = False):
     """Testing response from show.retrieve_all"""
-    show_info = show.retrieve_all(database_connection)
+    show_info = info.retrieve_all(database_connection)
     assert show_info is not None
     if print_response:
         print(json.dumps(show_info, indent=2))
@@ -177,7 +177,7 @@ def test_retrieve_all(database_connection: mysql.connector.connect,
 def test_retrieve_recent(database_connection: mysql.connector.connect,
                          print_response: bool = False):
     """Testing response from show.retrieve_recent"""
-    show_info = show.retrieve_recent(database_connection)
+    show_info = info.retrieve_recent(database_connection)
     assert show_info is not None
     if print_response:
         print(json.dumps(show_info, indent=2))
@@ -186,7 +186,7 @@ def test_retrieve_details_by_id(show_id: int,
                                 database_connection: mysql.connector.connect,
                                 print_response: bool = False):
     """Testing response from show.retrieve_details_by_id"""
-    show_details = show.retrieve_details_by_id(show_id, database_connection)
+    show_details = details.retrieve_by_id(show_id, database_connection)
     assert show_details is not None
     assert "host" in show_details
     assert "scorekeeper" in show_details
@@ -202,7 +202,7 @@ def test_retrieve_details_by_invalid_id(show_id: int,
                                         database_connection: mysql.connector.connect,
                                         print_response: bool = False):
     """Testing response from show.retrieve_details_by_id with invalid ID"""
-    show_details = show.retrieve_details_by_id(show_id, database_connection)
+    show_details = details.retrieve_by_id(show_id, database_connection)
     assert show_details is None
     if print_response:
         print(json.dumps(show_details, indent=2))
@@ -213,10 +213,10 @@ def test_retrieve_details_by_date(show_year: int,
                                   database_connection: mysql.connector.connect,
                                   print_response: bool = False):
     """Testing response from show.retrieve_details_by_date"""
-    show_details = show.retrieve_details_by_date(show_year,
-                                                 show_month,
-                                                 show_day,
-                                                 database_connection)
+    show_details = details.retrieve_by_date(show_year,
+                                            show_month,
+                                            show_day,
+                                            database_connection)
     assert show_details is not None
     assert "host" in show_details
     assert "scorekeeper" in show_details
@@ -234,10 +234,10 @@ def test_retrieve_details_by_invalid_date(show_year: int,
                                           database_connection: mysql.connector.connect,
                                           print_response: bool = False):
     """Testing repsonse from show.retrieve_details_by_date with incorrect date"""
-    show_details = show.retrieve_details_by_date(show_year,
-                                                 show_month,
-                                                 show_day,
-                                                 database_connection)
+    show_details = details.retrieve_by_date(show_year,
+                                            show_month,
+                                            show_day,
+                                            database_connection)
     assert show_details is None
     if print_response:
         print(json.dumps(show_details, indent=2))
@@ -246,8 +246,8 @@ def test_retrieve_details_by_date_string(show_date: str,
                                          database_connection: mysql.connector.connect,
                                          print_response: bool = False):
     """Testing repsonse from show.retrieve_details_by_date_string"""
-    show_details = show.retrieve_details_by_date_string(show_date,
-                                                        database_connection)
+    show_details = details.retrieve_by_date_string(show_date,
+                                                   database_connection)
     assert show_details is not None
     assert "host" in show_details
     assert "scorekeeper" in show_details
@@ -263,8 +263,8 @@ def test_retrieve_details_by_invalid_date_string(show_date: str,
                                                  database_connection: mysql.connector.connect,
                                                  print_response: bool = False):
     """Testing response from show.retrieve_details_by_date_string with invalud date string"""
-    show_details = show.retrieve_details_by_date_string(show_date,
-                                                        database_connection)
+    show_details = details.retrieve_by_date_string(show_date,
+                                                   database_connection)
     assert show_details is None
     if print_response:
         print(json.dumps(show_details, indent=2))
@@ -273,8 +273,8 @@ def test_retrieve_details_by_year(show_year: int,
                                   database_connection: mysql.connector.connect,
                                   print_response: bool = False):
     """Testing response from show.retrieve_details_by_year"""
-    show_details = show.retrieve_details_by_year(show_year,
-                                                 database_connection)
+    show_details = details.retrieve_by_year(show_year,
+                                            database_connection)
     assert show_details is not None
     if print_response:
         print(json.dumps(show_details, indent=2))
@@ -284,9 +284,9 @@ def test_retrieve_details_by_year_month(show_year: int,
                                         database_connection: mysql.connector.connect,
                                         print_response: bool = False):
     """Testing response from show.retrieve_details_by_year_month"""
-    show_details = show.retrieve_details_by_year_month(show_year,
-                                                       show_month,
-                                                       database_connection)
+    show_details = details.retrieve_by_year_month(show_year,
+                                                  show_month,
+                                                  database_connection)
     assert show_details is not None
     if print_response:
         print(json.dumps(show_details, indent=2))
@@ -294,7 +294,7 @@ def test_retrieve_details_by_year_month(show_year: int,
 def test_retrieve_all_details(database_connection: mysql.connector.connect,
                               print_response: bool = False):
     """Testing response from show.retrieve_all_show_details"""
-    show_details = show.retrieve_all_details(database_connection)
+    show_details = details.retrieve_all(database_connection)
     assert show_details is not None
     if print_response:
         print(json.dumps(show_details, indent=2))
@@ -302,7 +302,7 @@ def test_retrieve_all_details(database_connection: mysql.connector.connect,
 def test_retrieve_recent_details(database_connection: mysql.connector.connect,
                                  print_response: bool = False):
     """Testing response from show.retrieve_recent_show_details"""
-    show_details = show.retrieve_recent_details(database_connection)
+    show_details = details.retrieve_recent(database_connection)
     assert show_details is not None
     if print_response:
         print(json.dumps(show_details, indent=2))
