+++ wwdtm/__init__.py
@@ -5,4 +5,4 @@
 
 from wwdtm import guest, host, location, panelist, scorekeeper, show
 
-VERSION = "0.9.0.5"
+VERSION = "0.9.3.0"
